Generated code for return, load, store & call instrs
The hasResult function in register_allocation.cpp had a bug where it grouped all call variables as not having a result. This commit fixes the bug by checking if a call function has a result, such as the read function in minic which returns an int.

In addition to the bug fix, this commit also adds assembly code generation for return, load, store and call instructions. Helper functions were created where necessary and the code was tested as it was developed.