Refactor assembly codegen for Add/Sub/Mul/Icmp
This commit refactors the assembly code generation for the Add/Sub/Mul/Icmp opcodes. I will create a helper function for the assembly code generation switch cases and clean up the code after testing everything works correctly.

In addition, I added the EAX register to the Register enum and updated the getRegisterName function to support it.